requirements.txt
pandas==2.3.2
elasticsearch[async]==9.1.0
asyncpg==0.30.0
requests==2.32.5
fastapi==0.116.1
pydantic==2.11.7
Uvicorn==0.35.0
sqlalchemy[asyncio]==2.0.43
aiohttp==3.12.15
PyJWT==2.10.1
starlette==0.47.3